+++ src/index.js
@@ -4,6 +4,7 @@ import App from './App';
 import { Premise } from './shapes/Premise';
 import { Cut } from './shapes/Cut.js'
 import  _  from 'lodash'
+import  { handleCollisions } from './collisions.js'
 
 console.log("Starting...");
 
@@ -117,86 +118,7 @@ paper.on('cell:pointerdown', function(cellView, evt, x, y) {
 paper.on('cell:pointerup', function(cellView, evt, x, y) {
 
     let cell = cellView.model;
-    console.log(cell)
-    let cellsBelow = findModelsBelow(cell)
-    console.log("cellsBelow", cellsBelow)
+    
+    handleCollisions(cell)
 
-
-    if (cellsBelow.length) {
-        // Note that the findViewsFromPoint() returns the view for the `cell` itself.
-        let cellViewBelow = _.find(cellsBelow, function(c) { return c.id !== cell.id });
-        console.log("cellViewBelow", cellViewBelow)
-        // Prevent recursive embedding.
-        if (cellViewBelow && cellViewBelow.get('parent') !== cell.id) {
-            console.log("no recursive embedding detected")
-            let model = cellViewBelow
-            if (cell.attributes.type === "dia.Element.Premise") {
-                console.log("cell is cut", cell)
-                model.embed(cell);
-                cell.toFront();
-            } else if (cell.attributes.type === "dia.Element.Cut") {
-                console.log("cell is cut", cell);
-                console.log("hi")
-                model.toFront()
-                cell.embed(model)
-                model.toFront()
-            } else {
-                console.error("Invalid cell type detected!")
-            }
-        }
-    }
 });
-
-function findModelsBelow(element) {
-    let elementbbox = {
-        width: element.attributes.attrs.rect.width,
-        height: element.attributes.attrs.rect.height,
-        x: element.attributes.position.x,
-        y: element.attributes.position.y
-    }
-    let cells = graph.getCells()
-    let collisions = []
-    console.log("cells", cells)
-    for (const cell of cells) {
-        if (cell.id === element.id){
-            continue
-        }
-        let otherbbox = {
-            width: cell.attributes.attrs.rect.width,
-            height: cell.attributes.attrs.rect.height,
-            x: cell.attributes.position.x,
-            y: cell.attributes.position.y
-        }
-        console.log("bbox", elementbbox)
-        console.log("otherbbox", otherbbox)
-        if (contains(elementbbox,otherbbox)) {
-            console.log("contains!", cell)
-            collisions.push(cell)
-        }
-    }
-    console.log("colisions", collisions)
-    return collisions
-}
-
-function contains(bbox, otherbbox) {
-    let bbox_info = {
-        left_x: bbox.x,
-        right_x: bbox.x + bbox.width,
-        top_y: bbox.y,
-        bottom_y: bbox.y + bbox.height
-    }
-    let otherbbox_info = {
-        left_x: otherbbox.x,
-        right_x: otherbbox.x + otherbbox.width,
-        top_y: otherbbox.y,
-        bottom_y: otherbbox.y + otherbbox.height
-    }
-    console.log(bbox_info, otherbbox_info)
-    if (bbox_info.left_x < otherbbox_info.left_x && bbox_info.right_x > otherbbox_info.right_x && bbox_info.top_y < otherbbox_info.top_y && bbox_info.bottom_y > otherbbox_info.bottom_y) {
-        console.log("bbox contains otherbbox", bbox, otherbbox);
-        return true;
-    } else {
-        console.log("nop")
-        return false;
-    }
-}
